Refactor functional test, separate it into base for test, anonymous user test & logged in user test
